+++ crates/cli/src/commands/activate.rs
@@ -2,6 +2,7 @@ use crate::session::ProtoSession;
 use crate::workflows::{ExecWorkflow, ExecWorkflowParams};
 use clap::Args;
 use indexmap::IndexMap;
+use miette::IntoDiagnostic;
 use proto_core::{Id, PROTO_PLUGIN_KEY, ToolContext, UnresolvedVersionSpec};
 use serde::Serialize;
 use starbase::AppResult;
@@ -61,13 +62,6 @@ pub async fn activate(session: ProtoSession, args: ActivateArgs) -> AppResult {
     // Load necessary tools so that we can extract info
     let mut workflow = ExecWorkflow::new(session.load_tools().await?, config);
 
-    // Put shims first so that they can detect newer versions
-    if !args.no_shim {
-        workflow
-            .paths
-            .push_front(session.env.store.shims_dir.clone());
-    }
-
     // Aggregate our environment/shell exports
     workflow
         .prepare_environment(
@@ -112,7 +106,12 @@ pub async fn activate(session: ProtoSession, args: ActivateArgs) -> AppResult {
         workflow.env.insert("PROTO_VERSION".into(), None);
     }
 
-    // Put bins last as a last resort lookup
+    if !args.no_shim {
+        workflow
+            .paths
+            .push_back(session.env.store.shims_dir.clone());
+    }
+
     if !args.no_bin {
         workflow.paths.push_back(session.env.store.bin_dir.clone());
     }
@@ -227,11 +226,15 @@ fn print_activation_exports(
 
     // Set new `PATH`
     if !workflow.paths.is_empty() {
-        let joined_paths = workflow.join_paths()?;
-
-        if let Some(path) = joined_paths.as_ref().and_then(|inner| inner.to_str()) {
-            output.push(shell.format_env_set("_PROTO_ACTIVATED_PATH", path));
-        }
+        output.push(
+            shell.format_env_set(
+                "_PROTO_ACTIVATED_PATH",
+                env::join_paths(&workflow.paths)
+                    .into_diagnostic()?
+                    .to_str()
+                    .unwrap_or_default(),
+            ),
+        );
 
         let paths = workflow
             .reset_paths(&session.env.store.dir)